+++ src/Extracting/Shared/ApiResourceResponseTools.php
@@ -9,52 +9,56 @@ use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Pagination\Paginator;
+use Illuminate\Support\Arr;
 use Knuckles\Camel\Extraction\ExtractedEndpointData;
 use Knuckles\Scribe\Tools\ConsoleOutputUtils as c;
 use Knuckles\Scribe\Tools\ErrorHandlingUtils as e;
 use Knuckles\Scribe\Tools\Utils;
+use Mpociot\Reflection\DocBlock;
+use Mpociot\Reflection\DocBlock\Tag;
+use ReflectionClass;
 
 class ApiResourceResponseTools
 {
     public static function fetch(
-        string $apiResourceClass, bool $isCollection, $modelInstantiator,
+        string $apiResourceClass, bool $isCollection, ?callable $modelInstantiator,
         ExtractedEndpointData $endpointData, array $pagination, array $additionalData
     )
     {
-        try {
-            $resource = ApiResourceResponseTools::getApiResourceOrCollectionInstance(
-                $apiResourceClass, $isCollection, $modelInstantiator, $pagination, $additionalData
-            );
-            $response = ApiResourceResponseTools::getApiResourceResponse($resource, $endpointData);
-            return $response->getContent();
-        } catch (Exception $e) {
-            c::warn('Exception thrown when fetching Eloquent API resource response for ' . $endpointData->name());
-            e::dumpExceptionIfVerbose($e);
-
-            return null;
-        }
+        $resource = static::getApiResourceOrCollectionInstance(
+            $apiResourceClass, $isCollection, $modelInstantiator, $pagination, $additionalData
+        );
+        $response = static::callApiResourceAndGetResponse($resource, $endpointData);
+        return $response->getContent();
     }
 
-    public static function getApiResourceResponse(JsonResource $resource, ExtractedEndpointData $endpointData): JsonResponse
+    public static function callApiResourceAndGetResponse(JsonResource $resource, ExtractedEndpointData $endpointData): JsonResponse
     {
         $uri = Utils::getUrlWithBoundParameters($endpointData->route->uri(), $endpointData->cleanUrlParameters);
         $method = $endpointData->route->methods()[0];
         $request = Request::create($uri, $method);
         $request->headers->add(['Accept' => 'application/json']);
+        // Set the route properly, so it works for users who have code that checks for the route.
+        $request->setRouteResolver(fn() => $endpointData->route);
+
+        $previousBoundRequest = app('request');
         app()->bind('request', fn() => $request);
 
-        // Set the route properly, so it works for users who have code that checks for the route.
-        return $resource->toResponse(
-            $request->setRouteResolver(fn() => $endpointData->route)
-        );
+        $response = $resource->toResponse($request);
+
+        app()->bind('request', fn() => $previousBoundRequest);
+
+        return $response;
     }
 
     public static function getApiResourceOrCollectionInstance(
-        string $apiResourceClass, bool $isCollection, $modelInstantiator,
-        array  $paginationStrategy = [], array $additionalData = []
+        string $apiResourceClass, bool $isCollection, ?callable $modelInstantiator,
+        array $paginationStrategy = [], array $additionalData = []
     ): JsonResource
     {
-        $modelInstance = $modelInstantiator();
+        // If the API Resource uses an empty $resource (e.g. an empty array), the $modelInstantiator will be null
+        // See https://github.com/knuckleswtf/scribe/issues/652
+        $modelInstance = is_callable($modelInstantiator) ? $modelInstantiator() : [];
         try {
             $resource = new $apiResourceClass($modelInstance);
         } catch (Exception) {
@@ -91,5 +95,25 @@ class ApiResourceResponseTools
         }
 
         return $resource->additional($additionalData);
+    }
+
+    /**
+     * Check if the ApiResource class has an `@mixin` docblock, and fetch the model from there.
+     */
+    public static function tryToInferApiResourceModel(string $apiResourceClass): string|null
+    {
+        $class = new ReflectionClass($apiResourceClass);
+        $docBlock = new DocBlock($class->getDocComment() ?: '');
+        /** @var Tag|null $mixinTag */
+        $mixinTag = Arr::first(Utils::filterDocBlockTags($docBlock->getTags(), 'mixin'));
+        if (empty($mixinTag) || empty($modelClass = trim($mixinTag->getContent()))) {
+            return null;
+        }
+
+        if (class_exists($modelClass)) {
+            return $modelClass;
+        }
+
+        return null;
     }
 }