+++ src/ScribeServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Knuckles\Scribe;
 
+use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\ServiceProvider;
 use Knuckles\Scribe\Commands\DiffConfig;
 use Knuckles\Scribe\Commands\GenerateDocumentation;
@@ -11,9 +12,12 @@ use Knuckles\Scribe\Matching\RouteMatcher;
 use Knuckles\Scribe\Matching\RouteMatcherInterface;
 use Knuckles\Scribe\Tools\BladeMarkdownEngine;
 use Knuckles\Scribe\Tools\Utils;
+use Knuckles\Scribe\Writing\CustomTranslationsLoader;
 
 class ScribeServiceProvider extends ServiceProvider
 {
+    public static bool $customTranslationLayerLoaded = false;
+
     public function boot()
     {
         $this->registerViews();
@@ -24,10 +28,7 @@ class ScribeServiceProvider extends ServiceProvider
 
         $this->registerCommands();
 
-        $this->loadJsonTranslationsFrom(__DIR__.'/../lang');
-        $this->publishes([
-            __DIR__.'/../lang' => $this->app->langPath('vendor/scribe'),
-        ], 'scribe-translations');
+        $this->configureTranslations();
 
         // Bind the route matcher implementation
         $this->app->bind(RouteMatcherInterface::class, config('scribe.routeMatcher', RouteMatcher::class));
@@ -50,6 +51,16 @@ class ScribeServiceProvider extends ServiceProvider
             $routesPath = Utils::isLumen() ? __DIR__ . '/../routes/lumen.php' : __DIR__ . '/../routes/laravel.php';
             $this->loadRoutesFrom($routesPath);
         }
+    }
+
+    protected function configureTranslations(): void
+    {
+        $this->publishes([
+            __DIR__.'/../lang/' => $this->app->langPath(),
+        ], 'scribe-translations');
+
+        $this->loadTranslationsFrom($this->app->langPath('scribe.php'), 'scribe');
+        $this->loadTranslationsFrom(realpath(__DIR__ . '/../lang'), 'scribe');
     }
 
     protected function registerViews(): void
@@ -94,5 +105,17 @@ class ScribeServiceProvider extends ServiceProvider
                 DiffConfig::class,
             ]);
         }
+    }
+
+    // Allows our custom translation layer to be loaded on demand,
+    // so we minimize issues with interference from framework/package/environment.
+    // ALso, Laravel's `app->runningInConsole()` isn't reliable enough. See issue #676
+    public function loadCustomTranslationLayer(): void
+    {
+        $this->app->extend('translation.loader', function ($defaultFileLoader) {
+            return new CustomTranslationsLoader($defaultFileLoader);
+        });
+        $this->app->forgetInstance('translator');
+        self::$customTranslationLayerLoaded = true;
     }
 }