+++ tests/Fixtures/TestEmptyApiResource.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace Knuckles\Scribe\Tests\Fixtures;
+
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class TestEmptyApiResource extends JsonResource
+{
+    /**
+     * Create a new resource instance.
+     *
+     * @param  mixed  $resource
+     * @return void
+     */
+    public function __construct($resource = [])
+    {
+        $this->resource = $resource;
+    }
+
+    /**
+     * Get any additional data that should be returned with the resource array.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return array
+     */
+    public function with($request)
+    {
+        return [
+            'request-id' => 'ea02ebc1-4e3c-497f-9ea8-7a1ac5008af2',
+            'error_code' => 0,
+            'messages'   => []
+        ];
+    }
+
+    /**
+     * Transform the resource into an array.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     *
+     * @return array
+     */
+    public function toArray($request)
+    {
+        return [];
+    }
+}